HS-1020 - rename form, add to layout
